add lite dual db tests
add full dual db tests

test EE8 features only

fix java 2 security errors
