--- dev/com.ibm.ws.transaction.cloudsim_fat/fat/src/com/ibm/ws/transaction/test/DualServerDynamicTest.java
@@ -1,22 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2019 IBM Corporation and others.
- * All rights reserved. This program and the accompanying materials
- * are made available under the terms of the Eclipse Public License v1.0
- * which accompanies this distribution, and is available at
- * http://www.eclipse.org/legal/epl-v10.html
- *
- * Contributors:
- *     IBM Corporation - initial API and implementation
- *******************************************************************************/
-package com.ibm.ws.transaction.test;
-
-import org.junit.runner.RunWith;
-
-import componenttest.custom.junit.runner.FATRunner;
-import componenttest.custom.junit.runner.Mode;
-
-@Mode
-@RunWith(FATRunner.class)
-public class DualServerDynamicTest extends DualServerDynamicTestBase {
-
-}
