+++ dev/com.ibm.ws.transaction.cloudsim_fat/fat/src/com/ibm/ws/transaction/test/FATSuite.java
@@ -25,17 +25,17 @@ import componenttest.topology.impl.LibertyServerFactory;
 @SuiteClasses({
                 SimpleFS2PCCloudTest.class,
                 Simple2PCCloudTest.class,
-                DualServerDynamicTest.class
+                DualServerDynamicFSTest.class,
+                DualServerDynamicDBTest.class
 })
 public class FATSuite {
 
     private static LibertyServer server1 = LibertyServerFactory.getLibertyServer("com.ibm.ws.transaction_FSCLOUD001");
     private static LibertyServer server2 = LibertyServerFactory.getLibertyServer("com.ibm.ws.transaction_FSCLOUD002");
-    // Using the RepeatTests @ClassRule will cause all tests to be run twice.
-    // First without any modifications, then again with all features upgraded to their EE8 equivalents.
-    @ClassRule
-    public static RepeatTests r = RepeatTests.withoutModification()
-                    .andWith(FeatureReplacementAction.EE8_FEATURES());
+    // We don't repeat these tests, when they run in full mode doubling the number of tests can mean that the whole
+    // suite can take longer than the 3 hour threshold for a FAT suite.
+    //
+    // So run the suite with the EE8 Feature set only.
 
     @BeforeClass
     public static void beforeSuite() throws Exception {
