+++ dev/com.ibm.ws.transaction.cloudsim_fat/publish/servers/com.ibm.ws.transaction_CLOUD002/bootstrap.properties
@@ -1,4 +1,5 @@
 bootstrap.include=../testports.properties
 com.ibm.ws.logging.trace.specification=com.ibm.tx.jta.util.TxTMHelper=all=enabled
 ds.loglevel=DEBUG
+com.ibm.tx.auditRecovery=true
 osgi.console=5471
