+++ dev/com.ibm.ws.transaction.cloudsim_fat/publish/servers/com.ibm.ws.transaction_FSCLOUD002/server.xml
@@ -11,12 +11,12 @@
 <server>
    
     <featureManager>
-      <feature>servlet-3.1</feature>
-      <feature>jdbc-4.1</feature>
+      <feature>servlet-4.0</feature>
+      <feature>jdbc-4.2</feature>
+      <feature>jndi-1.0</feature>
       <feature>componentTest-1.0</feature>
       <feature>txfat-1.0</feature>
-      <feature>osgiconsole-1.0</feature>
-      <feature>cdi-1.2</feature>
+      <feature>cdi-2.0</feature>
       <feature>timedexit-1.0</feature>
     </featureManager>
     
@@ -37,5 +37,8 @@
 	
     <javaPermission codebase="${server.config.dir}/derby/derby.jar" className="java.security.AllPermission"/>
     <javaPermission codebase="${server.config.dir}/apps/transaction.war" className="java.security.AllPermission"/>
+    <javaPermission className="java.lang.RuntimePermission" name="*" actions="*"/>
+    <javaPermission className="java.io.FilePermission" name="ALL FILES" actions="read,write,delete"/>
+    <javaPermission className="java.util.PropertyPermission"  name="*" actions="read" />
     <logging traceSpecification="Transaction=all:com.ibm.ws.jdbc.*=all:RRA=all:logservice=all=enabled:*=info=enabled"/>
 </server>
