+++ dev/com.ibm.ws.transaction.cloudsim_fat/test-applications/transaction/src/com/ibm/ws/transaction/web/Simple2PCCloudServlet.java
@@ -10,7 +10,6 @@
  *******************************************************************************/
 package com.ibm.ws.transaction.web;
 
-import java.io.IOException;
 import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -18,41 +17,18 @@ import java.sql.Statement;
 
 import javax.annotation.Resource;
 import javax.annotation.Resource.AuthenticationType;
-import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.sql.DataSource;
-import javax.transaction.Transaction;
-
-import com.ibm.tx.jta.ExtendedTransactionManager;
-import com.ibm.tx.jta.TransactionManagerFactory;
-import com.ibm.tx.jta.ut.util.XAResourceImpl;
-import com.ibm.ws.cloudtx.ut.util.LastingXAResourceImpl;
-
-import componenttest.app.FATServlet;
 
 @WebServlet("/Simple2PCCloudServlet")
-public class Simple2PCCloudServlet extends FATServlet {
-
-    /**  */
-    private static final String filter = "(testfilter=jon)";
+public class Simple2PCCloudServlet extends Base2PCCloudServlet {
 
     @Resource(name = "jdbc/derby", shareable = true, authenticationType = AuthenticationType.APPLICATION)
     DataSource ds;
     @Resource(name = "jdbc/tranlogDataSource", shareable = true, authenticationType = AuthenticationType.APPLICATION)
     DataSource dsTranLog;
-
-    @Override
-    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        XAResourceImpl.setStateFile(System.getenv("WLP_OUTPUT_DIR") + "/../shared/" + LastingXAResourceImpl.STATE_FILE_ROOT);
-        super.doGet(request, response);
-    }
-
-    public void commitSuicide(HttpServletRequest request,
-                              HttpServletResponse response) throws Exception {
-        Runtime.getRuntime().halt(0);
-    }
 
     public void testLeaseTableAccess(HttpServletRequest request,
                                      HttpServletResponse response) throws Exception {
@@ -88,43 +64,8 @@ public class Simple2PCCloudServlet extends FATServlet {
         }
     }
 
-    public void setupRec001(HttpServletRequest request,
-                            HttpServletResponse response) throws Exception {
-        final ExtendedTransactionManager tm = TransactionManagerFactory
-                        .getTransactionManager();
-        System.out.println("NYTRACE: setUpRec001 started for tm: " + tm);
-        try {
-            tm.begin();
-
-            final Transaction tx = tm.getTransaction();
-
-            System.out.println("NYTRACE: setUpRec001 got transaction tx: " + tx);
-            LastingXAResourceImpl xares1 = new LastingXAResourceImpl();
-            System.out.println("NYTRACE: setUpRec001 working with res: " + xares1);
-            xares1.setCommitAction(XAResourceImpl.DIE);
-            try {
-                tx.enlistResource(xares1);
-            } catch (IllegalStateException ex) {
-                System.out.println("NYTRACE: setUpRec001 caught: " + ex);
-                // Assume transient (recovery incomplete) and retry after pause
-                Thread.sleep(5000);
-                tx.enlistResource(xares1);
-            }
-
-            tx.enlistResource(new LastingXAResourceImpl());
-            tx.enlistResource(new LastingXAResourceImpl());
-
-            tm.commit();
-        } catch (Exception e) {
-            System.out.println("NYTRACE: ImplodeServlet caught exc: " + e);
-            e.printStackTrace();
-        }
-    }
-
     public void modifyLeaseOwner(HttpServletRequest request,
                                  HttpServletResponse response) throws Exception {
-//        InitialContext context = new InitialContext();
-//        DataSource dsTranLog = (DataSource) context.lookup("java:comp/env/jdbc/tranlogDataSource");
 
         Connection con = dsTranLog.getConnection();
         try {
@@ -157,4 +98,5 @@ public class Simple2PCCloudServlet extends FATServlet {
             System.out.println("modifyLeaseOwner: caught exception in testSetup: " + ex);
         }
     }
+
 }
